+++ sql_queries/statements
@@ -1,0 +1,27 @@
+-- Feature 3 : Story 1 - Loan Account Table Structure
+-----------------------------------------------------
+CREATE TABLE `loan_data` (
+  `FIRST_NAME` varchar(45) DEFAULT NULL,
+  `LAST_NAME` varchar(45) DEFAULT NULL,
+  `FATHER_NAME` varchar(45) DEFAULT NULL,
+  `AGE` varchar(45) DEFAULT NULL,
+  `EMAIL` varchar(45) DEFAULT NULL,
+  `PHONE_NUMBER` varchar(45) DEFAULT NULL,
+  `ALTERNATIVE_PHONE_NUMBER` varchar(45) DEFAULT NULL,
+  `ACCOUNT_TYPE` varchar(45) DEFAULT NULL,
+  `CURRENT_ADDRESS` varchar(45) DEFAULT NULL,
+  `PERMANENT_ADDRESS` varchar(45) DEFAULT NULL,
+  `ADHAAR_NUMBER` varchar(45) NOT NULL,
+  `PAN_NUMBER` varchar(45) NOT NULL,
+  `NOMINEE_NAME` varchar(45) DEFAULT NULL,
+  `NOMINEE_ADHAAR_NUMBER` varchar(45) NOT NULL,
+  `NOMINEE_ACCOUNT_NUMBER` varchar(45) NOT NULL,
+  `ACCOUNT_NUMBER` varchar(45) NOT NULL,
+  `LOAN_ACCOUNT_NUMBER` varchar(45) NOT NULL,
+  `LOAN_TYPE` varchar(45) DEFAULT NULL,
+  `ASSETS` varchar(45) DEFAULT NULL,
+  `HOW_MUCH_LOAN_REQUIRED` varchar(45) DEFAULT NULL,
+  `APPROVAL` varchar(45) DEFAULT NULL,
+  PRIMARY KEY (`ADHAAR_NUMBER`,`PAN_NUMBER`,`NOMINEE_ADHAAR_NUMBER`,`NOMINEE_ACCOUNT_NUMBER`,`ACCOUNT_NUMBER`,`LOAN_ACCOUNT_NUMBER`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
+------------------------------------------------------------------------------------------------------------------------------------
