+++ src/main/webapp/WEB-INF/views/apply.jsp
@@ -1,5 +1,5 @@
 <%@page import="java.util.Objects"%>
-<%@page import="com.khadri.spring.mvc.form.CustomerForm"%>
+<%@page import="com.khadri.spring.mvc.loan.form.CustomerForm"%>
 <%@ page language="java" contentType="text/html; charset=ISO-8859-1"
 	pageEncoding="ISO-8859-1"%>
 <!DOCTYPE html>
