--- src/main/webapp/a.jsp
@@ -1,23 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html>
-<head>
-<style type="text/css">
-.center {
-  padding: 70px 0;
-  text-align: center;
-}
-</style>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-</head>
-<body>
-
-	<div class="center">
-		<div>first</div>
-		<div>second</div>
-	</div>
-
-</body>
-</html>
